move some component to ui dir
