+++ src/components/CopyButton/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '../Button';
+import Button from '../ui/Button';
 import { CheckMarkIcon, CopyIcon } from '../BlockWrap/icon';
 import { useConfig } from '../../context/ConfigProvider';
