+++ src/index.less
@@ -385,3 +385,16 @@ ds-markdown-math {
 .ds-button:hover {
   background-color: var(--ds-button-hover-color);
 }
+
+[id^='dmermaid-'] {
+  height: 0;
+  overflow: hidden;
+}
+
+.ds-button.ds-icon-button {
+  padding: 0 4px;
+}
+
+.ds-button.ds-icon-button .ds-button__icon {
+  margin-right: 0;
+}
